Requested changes in code style
All files listed have been changed with brunoocasali's review suggestions.
