+++ .rubocop.yml
@@ -1,3 +1,6 @@
+Style/FrozenStringLiteralComment:
+  Enabled: false
+
 Documentation:
   Enabled: false
