+++ Rakefile
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Add your own tasks in files placed in lib/tasks ending in .rake,
 # for example lib/tasks/capistrano.rake,
 # and they will automatically be available to Rake.
