+++ app/channels/application_cable/channel.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 module ApplicationCable
   class Channel < ActionCable::Channel::Base
   end
