+++ app/channels/application_cable/connection.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 module ApplicationCable
   class Connection < ActionCable::Connection::Base
   end
