+++ app/controllers/home_controller.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class HomeController < ApplicationController
   before_action :authenticate_user!
   skip_before_action :verify_authenticity_token
