+++ app/controllers/user_controller.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class UserController < ApplicationController
   def user_home
     redirect_to graphic_path if user_signed_in?
