+++ app/jobs/application_job.rb
@@ -1,4 +1,2 @@
-# frozen_string_literal: true
-
 class ApplicationJob < ActiveJob::Base
 end
