+++ app/models/application_record.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
 end
