+++ app/models/contract.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class Contract < ApplicationRecord
   belongs_to :user
   belongs_to :report
