+++ app/models/report.rb
@@ -1,4 +1,2 @@
-# frozen_string_literal: true
-
 class Report < ApplicationRecord
 end
