+++ app/views/home/spreadsheet.html.slim
@@ -2,12 +2,10 @@ html
   head
     meta charset="utf-8" /
     title Sales Up - Spreadsheet
-    meta content="app, web app, responsive, responsive layout, admin, admin panel, admin dashboard, flat, flat ui, ui kit, AngularJS, ui route, charts, widgets, components" name="description" /
-   
     = stylesheet_link_tag "application", media: 'all'
   body
     .app
-      #content role="main" 
+      #content role="main"
         .app-content-body
           .form-inline.row.menu.panel.panel-default
             .form-group#menu-logo
@@ -21,7 +19,7 @@ html
 
             .form-group#menu-search
               form.form-group method="post" action="/search"
-                select#monthBox.form-control name="month" 
+                select#monthBox.form-control name="month"
                   -for report in Report.all.order('month_numb', 'year').reverse
                     option value="#{report.month}/#{report.year}" = report.month + "/" + report.year.to_s
                 button.btn.btn-primary.btn-addon.btn-lg.margin name="where" value="spreadsheet" type="submit" Search
@@ -42,14 +40,14 @@ html
               h2.h2.font-thin.panel-heading
                 = $current_report.report_name + " --- Goal value: R$" + $current_report.goal.to_s + ")"
               div
-                table.table ui-jq="footable" 
+                table.table ui-jq="footable"
                   thead
                     tr
                       th data-breakpoints="xs"  Day
                       th Weekdays
                       th Daily Goal
                       th Contract Sum
-                      th data-breakpoints="xs"  Store Name 
+                      th data-breakpoints="xs"  Store Name
                       th data-breakpoints=("xs sm")  Contract Values
                       th data-breakpoints=("xs sm md") Salesman Name
                   tbody
