+++ app/views/management/manager.html.slim
@@ -2,8 +2,6 @@ html
   head
     meta charset="utf-8" /
     title Sales Up - Management
-    meta content="app, web app, responsive, responsive layout, admin, admin panel, admin dashboard, flat, flat ui, ui kit, AngularJS, ui route, charts, widgets, components" name="description" /
-
     = stylesheet_link_tag "application", media: 'all'
   body
     .app
