+++ app/views/user/forgot_password.html.slim
@@ -2,24 +2,21 @@ html
   head
     meta charset="utf-8" /
     title Sales Up - Forgot your password
-
-    meta content="app, web app, responsive, responsive layout, admin, admin panel, admin dashboard, flat, flat ui, ui kit, AngularJS, ui route, charts, widgets, components" name="description" /
-
     = stylesheet_link_tag "application", media: 'all'
   body
     .app.app-header-fixed
-      .container.w-xxl.w-auto-xs ng-init=("app.settings.container = false;") 
+      .container.w-xxl.w-auto-xs ng-init=("app.settings.container = false;")
 
         .m-b-lg
         .text-center
-		  a href="http://site.trustvox.com.br/" 
+		  a href="http://site.trustvox.com.br/"
 		    img src="/assets/img/trustvox-login-logo.png" /
 		  h1 Sales Up
 		  h4 Business Performance Manager
 
           .wrapper.text-center
             strong Enter your email to reset your password
-          form name="reset" ng-init="isCollapsed=true" 
+          form name="reset" ng-init="isCollapsed=true"
             .list-group.list-group-sm
               .list-group-item
                 input.form-control.no-border ng-model="email" placeholder="Email" required="" type="email" /
