+++ app/views/user/user_home.html.slim
@@ -2,16 +2,13 @@ html
   head
     meta charset="utf-8" /
     title Sales Up - Login
-
-    meta content="app, web app, responsive, responsive layout, admin, admin panel, admin dashboard, flat, flat ui, ui kit, AngularJS, ui route, charts, widgets, components" name="description" /
-
     = stylesheet_link_tag "application", media: 'all'
   body
     .app.app-header-fixed
-      .container.w-xxl.w-auto-xs ng-controller="SigninFormController" ng-init=("app.settings.container = false;") 
+      .container.w-xxl.w-auto-xs ng-controller="SigninFormController" ng-init=("app.settings.container = false;")
         .m-b-lg
         .text-center
-		  a href="http://site.trustvox.com.br/" 
+		  a href="http://site.trustvox.com.br/"
 		    img src="/assets/img/trustvox-login-logo.png" /
 		  h1 Sales Up
 		  h4 Business Performance Manager
@@ -28,4 +25,4 @@ html
             = f.submit "Log in", :class => "btn btn-lg btn-primary btn-block", :where => "graphic"
             .text-center.m-t.m-b
               = link_to 'Forgot your password?', forgot_password_path
- 
+
