+++ config.ru
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # This file is used by Rack-based servers to start the application.
 
 require_relative 'config/environment'
