+++ config/boot.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../Gemfile', __dir__)
 
 require 'bundler/setup' # Set up gems listed in the Gemfile.
