+++ config/environment.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Load the Rails application.
 require_relative 'application'
