+++ config/environments/production.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 Rails.application.configure do
   # Settings specified here will take
   # precedence over those in config/application.rb.
