+++ config/puma.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Puma can serve each request in a thread from an internal thread pool.
 # The `threads` method setting takes two numbers: a minimum and maximum.
 # Any libraries that use thread pools should be configured to match
