+++ config/spring.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 %w[
   .ruby-version
   .rbenv-vars
