+++ spec/spec_helper.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
